requirements.txt
pandas==2.2.2
numpy>=1.24,<2
scipy>=1.10,<1.12
scikit-learn==1.4.2
xgboost==2.1.1
chembl_webresource_client==0.10.9
rdkit-pypi==2022.9.5
joblib==1.4.2
requests==2.32.3
tqdm==4.66.5
matplotlib==3.8.2
